lmb-796: make donwloadLink a getter and pass on BOI id to the component
