+++ app/templates/organen/orgaan/mandatarissen.hbs
@@ -30,6 +30,7 @@
       @bestuursperiode={{@model.selectedBestuursperiode}}
       @activeOnly={{this.activeOnly}}
       @bestuursorgaan={{this.model.bestuursorgaan}}
+      @bestuursorgaanInTijdId={{this.model.bestuursorgaanInTijdId}}
       @sort={{this.sort}}
     />
     <Shared::Tooltip
